Ifdef guards added. Comparison to the GTOC7 state that was read in has been added for easy comparison. Added variable that controls how many days you propagate in a phase.
git-svn-id: svn://svn.code.sf.net/p/emtg/emtg-goddard@204 77bb456d-b9d8-4424-8c4a-7c450a671a84

+++ branch/EMTG-GTOC7/Utilities/CheckYoSelf/src/orbitprop.cpp
@@ -20,7 +20,7 @@ void orbitprop(Body * body, const double & delta_t, const double & mu_sun)
 	coe2cartesian(body, mu_sun);
 }
 
-void orbitprop(Spacecraft * probe, const double & delta_t, const double & phase, const double & timestep, const double & mu_sun)
+void orbitprop(Spacecraft * probe, const double & delta_t, const int & phase, const int & timestep, const double & mu_sun)
 {
 	probe->M[phase][timestep] = probe->M[phase][timestep] + sqrt(mu_sun / ((probe->a[phase][timestep])*(probe->a[phase][timestep])*(probe->a[phase][timestep])))*delta_t;
